+++ R/api.R
@@ -37,7 +37,7 @@
 #'
 #' @export
 get_app <- function(request_params = list(),
-                    gallery = getOption("alteryx_gallery")) {
+                    gallery = get_gallery()) {
   endpoint <- "/api/v1/workflows/subscription/"
 
   content <- submit_get_request(gallery,
@@ -64,7 +64,7 @@ get_app <- function(request_params = list(),
 #' @export
 download_app <- function(app,
                          destfile,
-                         gallery = getOption("alteryx_gallery")) {
+                         gallery = get_gallery()) {
   class_check <- check_class(app, "app")
 
   endpoint <- "/api/v1/{appId}/package/"
@@ -110,7 +110,7 @@ download_app <- function(app,
 #'
 #' @export
 get_app_questions <- function(app,
-                              gallery = getOption("alteryx_gallery")) {
+                              gallery = get_gallery()) {
   class_check <- check_class(app, "app")
 
   request_params <- list()
@@ -173,7 +173,7 @@ NULL
 #' @export
 get_app_jobs <- function(app,
                          request_params = list(),
-                         gallery = getOption("alteryx_gallery")) {
+                         gallery = get_gallery()) {
   class_check <- check_class(app, "app")
 
   endpoint <- "/api/v1/workflows/{appId}/jobs/"
@@ -187,7 +187,7 @@ get_app_jobs <- function(app,
   job_ids <- lapply(content, function(x) {x$id})
   get_job_by_id <- function(job_id,
                             app,
-                            gallery = getOption("alteryx_gallery")) {
+                            gallery = get_gallery()) {
     request_params <- list()
     endpoint <- "/api/v1/jobs/{jobId}/"
     endpoint <- gsub("\\{jobId\\}", job_id, endpoint)
@@ -209,7 +209,7 @@ get_app_jobs <- function(app,
 #'
 #' @export
 get_job <- function(job,
-                    gallery = getOption("alteryx_gallery")) {
+                    gallery = get_gallery()) {
   class_check <- check_class(job, "job")
 
   request_params <- list()
@@ -220,6 +220,7 @@ get_job <- function(job,
   content <- submit_get_request(gallery,
                                 endpoint,
                                 request_params)
+  content$workerTag<-ifelse(is.null(content$workerTag),"",content$workerTag)
 
   content <- as.alteryx_job(content, job)
 
@@ -253,7 +254,7 @@ get_job <- function(job,
 #'
 #' @export
 get_job_output <- function(job,
-                           gallery = getOption("alteryx_gallery"),
+                           gallery = get_gallery(),
                            quiet = FALSE) {
   class_check <- check_class(job, "job")
 
@@ -323,6 +324,12 @@ get_job_output <- function(job,
 #' @param app A single \code{alteryx_app} returned from \code{get_app}
 #' @param answers Answers to required \code{app} questions created using
 #' \code{build_answers}
+#' @param priority Assign a priority level to jobs to control which jobs are
+#' run by each worker, or to reserve specific workers for higher priority
+#' requests. When running a workflow, users (if enabled by the Server Admin)
+#' can select a priority level of 'low', 'medium', 'high', or 'critical' to
+#' ensure certain jobs always take priority over others. If multiple jobs are
+#' queued, jobs run in priority order starting with the highest priority.
 #' @param track_job If \code{TRUE} this function will not return a value until
 #' the job completes on Alteryx Gallery
 #' @param sleep Amount of time to wait between job polls. Ignored if
@@ -351,11 +358,18 @@ NULL
 #' @export
 queue_job <- function(app,
                       answers,
+                      priority = "low",
                       track_job = FALSE,
                       sleep = 10,
                       timeout = 3600,
-                      gallery = getOption("alteryx_gallery")) {
+                      gallery = get_gallery()) {
+
   class_check <- check_class(app, "app")
+
+  priority <- get_priority(priority)
+  answers <- jsonlite::fromJSON(answers)
+  answers["priority"] <- priority
+  answers <- jsonlite::toJSON(answers, auto_unbox = TRUE)
 
   request_params <- list()
   app_id <- app$id